+++ api/src/telebot/telebot.cpp
@@ -36,7 +36,7 @@ bool init() {
 
         void onLinkRequest(stun::Client* client, const std::string& from) {}
 
-        void onLinkAccepted(stun::Client* client, const std::string& name, const std::string& ip, uint16_t port) {}
+        void onLinkAccepted(stun::Client* client, const std::string& name, const std::string& ip, int port) {}
 
         void onLinkDeclined(stun::Client* client, const std::string& name, const std::string& message) {}
     };
